@@ -21,20 +21,26 @@
   },
   "homepage": "https://github.com/mattdot/botauth#readme",
   "dependencies": {
+    "@types/cookie": "^0.3.0",
+    "@types/express": "^4.0.35",
+    "@types/passport": "^0.3.3",
+    "@types/passport-strategy": "^0.2.30",
+    "@types/winston": "^2.3.0",
     "botbuilder": "^3.5.1",
     "cookie": "^0.3.1",
     "passport": "^0.3.2",
-    "passport-strategy": "^1.0.0"
+    "passport-strategy": "^1.0.0",
+    "winston": "^2.3.1"
   },
   "devDependencies": {
     "chai": "^3.5.0",
     "gulp": "^3.9.1",
     "gulp-run": "^1.7.1",
-    "gulp-typescript": "^3.1.3",
-    "merge-stream": "^1.0.0",
-    "mocha": "^3.0.2",
+    "gulp-typescript": "^3.1.6",
+    "merge-stream": "^1.0.1",
+    "mocha": "^3.2.0",
     "tslint": "^3.15.1",
-    "typescript": "^2.0.3"
+    "typescript": "^2.2.2"
   },
   "typings": "lib/index",
   "files": [
