@@ -1,4 +1,5 @@
 "use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
 const url = require("url");
 const crypto = require("crypto");
 const builder = require("botbuilder");
@@ -83,7 +84,7 @@ class BotAuthenticator {
     }
     authenticate(providerId, options) {
         let authSteps = [
-                (session, args, skip) => {
+            (session, args, skip) => {
                 let user = this.profile(session, providerId);
                 if (user) {
                     skip({ response: (args || {}).response, resumed: builder.ResumeReason.forward });
@@ -99,7 +100,7 @@ class BotAuthenticator {
                     });
                 }
             },
-                (session, args, skip) => {
+            (session, args, skip) => {
                 if (args) {
                     if (args.resumed === builder.ResumeReason.completed || args.resumed === builder.ResumeReason.forward) {
                         skip({ response: args.response, resumed: builder.ResumeReason.forward });
